remove permissions modal prop drilling
