+++ src/app/components/permissionsModal/PermissionList.tsx
@@ -1,30 +1,11 @@
-import { usePermissions, useUnqiuePrograms } from "@/app/hooks/reactQuery";
-import { Permission } from "@/app/types/Permission";
-import { Program } from "@/app/types/Program";
+import { usePermissions } from "@/app/hooks/reactQuery";
 import { sortPermissions } from "@/lib/permissionsUtils";
 import { AccordionGroup, List, ListItem, ListItemContent } from "@mui/joy";
 import { lazy } from "react";
 
 const PermissionAccordion = lazy(() => import("./PermissionAccordion"));
 
-export interface PermissionsProps {
-  onAddMajorEditPermission: (permission: Permission, major: Program) => void;
-  onRemoveMajorEditPermission: (permission: Permission, major: Program) => void;
-  onRemovePermissions: (permission: Permission) => void;
-  onUpdateMajorEditPermissionExpirationDate: (
-    permission: Permission,
-    major: Program,
-    expirationDate: Date,
-  ) => void;
-}
-
-export default function PermissionsList({
-  onAddMajorEditPermission,
-  onRemoveMajorEditPermission,
-  onRemovePermissions,
-  onUpdateMajorEditPermissionExpirationDate,
-}: PermissionsProps) {
-  const { data: programs } = useUnqiuePrograms();
+export default function PermissionsList() {
   const { data: permissions } = usePermissions(sortPermissions);
 
   return (
@@ -34,16 +15,7 @@ export default function PermissionsList({
           permissions.map((permission, index) => (
             <ListItem key={index}>
               <ListItemContent>
-                <PermissionAccordion
-                  permission={permission}
-                  programs={programs ?? []}
-                  onAddMajorEditPermission={onAddMajorEditPermission}
-                  onRemoveMajorEditPermission={onRemoveMajorEditPermission}
-                  onRemovePermissions={onRemovePermissions}
-                  onUpdateMajorEditPermissionExpirationDate={
-                    onUpdateMajorEditPermissionExpirationDate
-                  }
-                />
+                <PermissionAccordion permission={permission} />
               </ListItemContent>
             </ListItem>
           ))}
